backend/requirements.txt
fastapi
uvicorn[standard]
requests
influxdb-client
psycopg2-binary
pydantic
